backend/requirements.txt
fastapi>=0.115.6,<1.0
uvicorn[standard]>=0.34.0,<1.0
google-genai>=1.21.1,<2.0
google-cloud-texttospeech>=2.24.0,<3.0
google-adk>=1.0.0,<2.0
pydantic>=2.10.4,<3.0
python-dotenv>=1.0.1,<2.0
firebase-admin>=6.4.0,<7.0
google-cloud-storage>=2.18.0,<3.0
google-cloud-firestore>=2.19.0,<3.0
fpdf2>=2.7.0,<3.0
Pillow>=10.0.0,<12.0
requests>=2.31.0,<3.0
